pyproject.toml
[project]
name = "projectwise_client"
version = "0.1.0"
description = "MCP Client"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp[cli]>=1.10.1",
    "mem0ai>=0.1.114",
    "nest-asyncio>=1.6.0",
    "openai>=1.93.0",
    "python-dotenv>=1.1.1",
    "tiktoken>=0.9.0",
]
